Integrate messaging with new dev-server
